+++ packages/components/atoms/f-card/src/tenants/da-DK.js
@@ -1,4 +1,0 @@
-export default {
-    locale: 'da-DK',
-    cardTitle: 'I am a Card Component (DK)'
-};