+++ packages/components/atoms/f-card/src/tenants/en-AU.js
@@ -1,4 +1,0 @@
-export default {
-    locale: 'en-AU',
-    cardTitle: 'I am a Card Component (AU)'
-};