+++ packages/components/atoms/f-card/src/tenants/en-GB.js
@@ -1,4 +1,0 @@
-export default {
-    locale: 'en-GB',
-    cardTitle: 'I am a Card Component (GB)'
-};