+++ packages/components/atoms/f-card/src/tenants/en-NZ.js
@@ -1,4 +1,0 @@
-export default {
-    locale: 'en-NZ',
-    cardTitle: 'I am a Card Component (NZ)'
-};