+++ packages/components/organisms/f-checkout/test-utils/component-objects/f-checkout.component.js
@@ -1,6 +1,5 @@
-/* global browser, $ */
-
-import {
+const Page = require('@justeat/f-wdio-utils/src/page.object');
+const {
     CHECKOUT_COMPONENT,
     ORDER_TIME_DROPDOWN,
     ORDER_TIME_DROPDOWN_OPTIONS,
@@ -10,195 +9,286 @@ import {
     KNOB_CHECKOUT_DROPDOWN,
     KNOB_BUTTON,
     SWITCH_USER_LINK,
-    GUEST_CHECKOUT_LOGIN_BUTTON,
-    GUEST_CHECKOUT_HEADER
-} from './f-checkout-selectors';
+    GUEST_CHECKOUT_HEADER,
+    GUEST_CHECKOUT_LOGIN_BUTTON
+} = require('./f-checkout-selectors');
 
-const { doesElementExist } = require('../../../../../../test/utils/webdriverio-extensions')(browser);
+class Checkout extends Page {
+    get component () { return $(CHECKOUT_COMPONENT); }
 
-const checkoutComponent = () => $(CHECKOUT_COMPONENT);
-const guestCheckoutHeader = () => $(GUEST_CHECKOUT_HEADER);
+    get orderTimeDropdown () { return $(ORDER_TIME_DROPDOWN); }
 
-// Dropdown
+    get orderTimeDropdownOptions () { return $$(ORDER_TIME_DROPDOWN_OPTIONS); }
 
-const orderTimeDropdown = () => $(ORDER_TIME_DROPDOWN);
-const orderTimeDropdownOptions = () => $$(ORDER_TIME_DROPDOWN_OPTIONS);
-const knobCheckoutDropdown = () => $(KNOB_CHECKOUT_DROPDOWN);
+    get knobCheckoutDropdown () { return $(KNOB_CHECKOUT_DROPDOWN); }
 
-// Buttons
-const goToPaymentButton = () => $(GO_TO_PAYMENT_BUTTON);
-const knobButton = () => $(KNOB_BUTTON);
-const guestCheckoutLoginButton = () => $(GUEST_CHECKOUT_LOGIN_BUTTON);
+    get goToPaymentButton () { return $(GO_TO_PAYMENT_BUTTON); }
 
-// Form Fields
+    get knobButton () { return $(KNOB_BUTTON); }
 
-const userNoteInput = () => $(USER_NOTE_INPUT);
-const switchUserLink = () => $(SWITCH_USER_LINK);
+    get userNoteInput () { return $(USER_NOTE_INPUT); }
 
-const fields = {
-    firstName: {
-        input: () => $(FIELDS.firstName.input),
-        error: () => $(FIELDS.firstName.error)
-    },
-    lastName: {
-        input: () => $(FIELDS.lastName.input),
-        error: () => $(FIELDS.lastName.error)
-    },
-    emailAddress: {
-        input: () => $(FIELDS.emailAddress.input),
-        error: () => $(FIELDS.emailAddress.error)
-    },
-    mobileNumber: {
-        input: () => $(FIELDS.mobileNumber.input),
-        error: () => $(FIELDS.mobileNumber.error)
-    },
-    addressLine1: {
-        input: () => $(FIELDS.addressLine1.input),
-        error: () => $(FIELDS.addressLine1.error)
-    },
-    addressLine2: {
-        input: () => $(FIELDS.addressLine2.input),
-        error: ''
-    },
-    addressCity: {
-        input: () => $(FIELDS.addressCity.input),
-        error: () => $(FIELDS.addressCity.error)
-    },
-    addressPostcode: {
-        input: () => $(FIELDS.addressPostcode.input),
-        error: () => $(FIELDS.addressPostcode.error),
-        typeError: () => $$(FIELDS.addressPostcode.typeError)
-    },
-    userNote: {
-        input: () => $(FIELDS.userNote.input),
-        error: ''
+    get switchUserLink () { return $(SWITCH_USER_LINK); }
+
+    get guestCheckoutHeader () { return $(GUEST_CHECKOUT_HEADER); }
+
+    get guestCheckoutLoginButton () { return $(GUEST_CHECKOUT_LOGIN_BUTTON); }
+
+    fields = {
+        firstName: {
+            get input () { return $(FIELDS.firstName.input); },
+            get error () { return $(FIELDS.firstName.error); }
+        },
+        lastName: {
+            get input () { return $(FIELDS.lastName.input); },
+            get error () { return $(FIELDS.lastName.error); }
+        },
+        emailAddress: {
+            get input () { return $(FIELDS.emailAddress.input); },
+            get error () { return $(FIELDS.emailAddress.error); }
+        },
+        mobileNumber: {
+            get input () { return $(FIELDS.mobileNumber.input); },
+            get error () { return $(FIELDS.mobileNumber.error); }
+        },
+        addressLine1: {
+            get input () { return $(FIELDS.addressLine1.input); },
+            get error () { return $(FIELDS.addressLine1.error); }
+        },
+        addressLine2: {
+            get input () { return $(FIELDS.addressLine2.input); },
+            get error () { return $(FIELDS.addressLine2.error); }
+        },
+        addressCity: {
+            get input () { return $(FIELDS.addressCity.input); },
+            get error () { return $(FIELDS.addressCity.error); }
+        },
+        addressPostcode: {
+            get input () { return $(FIELDS.addressPostcode.input); },
+            get error () { return $(FIELDS.addressPostcode.error); },
+            get typeError () { return $(FIELDS.addressPostcode.typeError); }
+        },
+        userNote: {
+            get input () { return $(FIELDS.userNote.input); },
+            get error () { return ''; }
+        }
     }
-};
+    /**
+     * @description
+     * Sets the data for the checkout component.
+     *
+     * @param {Object} checkout
+     * @param {String} checkout.type The checkout type
+     * @param {String} checkout.isAuthenticated The checkout authentication
+     * @param {String} checkout.isValid The checkout validation
+     */
 
-/**
- * @description
- * Changes checkout page to reflect checkout method to either delivery or collection depending on index given.
- *
- * @param {string} method The collection type: either 'delivery' or 'collection'
- */
+    open (checkout) {
+        let authQueryString = '';
+        let createGuestUrl = '&knob-Create%20Guest%20Url=%2Fcreate-guest.json';
+        const checkoutUrl = checkout.isValid ? `&knob-Get%20Checkout%20Url=%2Fcheckout-${checkout.type}.json` : '';
+        const basketUrl = `&knob-Get%20Basket%20Url=%2Fget-basket-${checkout.type}.json`;
+        const placeOrderUrl = '&knob-Place%20Order%20Url=%2Fplace-order.json';
+        const paymentPageUrlPrefix = '&knob-Payment%20Page%20Url%20Prefix=%23%2Fpay';
 
-exports.changeCheckoutMethod = method => {
-    const file = `/checkout-${method}.json`;
-    knobButton().click();
-    knobCheckoutDropdown().selectByVisibleText(file);
-};
+        if (checkout.isAuthenticated) {
+            authQueryString = '&knob-Auth%20token=a';
+            createGuestUrl = '';
+        }
+
+        const url = `checkout-component${checkoutUrl}&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json${authQueryString}${basketUrl}${createGuestUrl}${placeOrderUrl}${paymentPageUrlPrefix}`;
+        super.openComponent('organism', url);
+    }
+
+    waitForComponent () {
+        super.waitForComponent(this.component);
+    }
+
+    isComponentDisplayed () {
+        return this.component.isDisplayed();
+    }
+
+    isFieldErrorDisplayed (fieldName) {
+        return this.fields[fieldName].error.isDisplayed();
+    }
+
+    isFieldDisplayed (fieldName) {
+        return this.fields[fieldName].input.isDisplayed();
+    }
+
+    isPostcodeTypeErrorDisplayed () {
+        return this.fields.addressPostcode.typeError.isDisplayed();
+    }
+
+    isOrderTimeDropdownDisplayed () {
+        return this.orderTimeDropdown.isDisplayed();
+    }
+
+    userNoteMaxCharacterCount () {
+        return this.userNoteInput.getAttribute('maxlength');
+    }
+
+    clickPaymentButton () {
+        return this.goToPaymentButton.click();
+    }
+
+    switchUserLinkIsDisplayed () {
+        return this.switchUserLink.isDisplayed();
+    }
+
+    isGuestCheckoutLoginButtonDisplayed () {
+        return this.guestCheckoutLoginButton.isDisplayed();
+    }
+
+    clickGuestCheckoutLoginButton () {
+        return this.guestCheckoutLoginButton.click();
+    }
+
+    isGuestCheckoutHeaderDisplayed () {
+        return this.guestCheckoutHeader.isDisplayed();
+    }
 
 
+    /**
+    * @description
+    * Inputs user details into the registration component and submits the form.
+    *
+    * @param {Object} userInfo
+    * @param {String} userInfo.firstName The user's first name
+    * @param {String} userInfo.lastName The user's last name
+    * @param {String} userInfo.email The user's e-mail address
+    * @param {String} userInfo.password The user's password
+    */
+    submitForm () {
+        this.fields.firstName.input.setValue(userInfo.firstName);
+        this.fields.lastName.input.setValue(userInfo.lastName);
+        this.fields.email.input.setValue(userInfo.email);
+        this.fields.password.input.setValue(userInfo.password);
+        this.createAccountButton.click();
+    }
 
-exports.isFieldErrorDisplayed = fieldName => fields[fieldName].error().isDisplayed();
-exports.isFieldDisplayed = fieldName => fields[fieldName].input().isDisplayed();
-exports.isPostCodeTypeErrorDisplayed = () => fields.addressPostcode.typeError().length > 0;
-exports.waitForCheckoutComponent = () => checkoutComponent().waitForExist();
-exports.isCheckoutComponentDisplayed = () => checkoutComponent().isDisplayed();
-exports.isOrderTimeDropdownDisplayed = () => orderTimeDropdown().isDisplayed();
-exports.userNoteMaxCharacterCount = () => userNoteInput().getAttribute('maxlength');
-exports.clickPaymentButton = () => goToPaymentButton().click();
-exports.switchUserLinkIsDisplayed = () => switchUserLink().isDisplayed();
-exports.isGuestCheckoutLoginButtonDisplayed = () => guestCheckoutLoginButton().isDisplayed();
-exports.isGuestCheckoutHeaderDisplayed = () => guestCheckoutHeader().isDisplayed();
-exports.clickGuestCheckoutLogin = () => guestCheckoutLoginButton().click();
+    /**
+    * @description
+    * Inputs address details into the checkout component.
+    *
+    * @param {Object} addressInfo
+    * @param {String} addressInfo.mobileNumber The user's mobile number
+    * @param {String} addressInfo.line1 First line of the user's address
+    * @param {String} addressInfo.line2 Second line of the user's address
+    * @param {String} addressInfo.city City of the user's address
+    * @param {String} addressInfo.postcode Postcode of the user's address
+    * @param {String} addressInfo.note The user's extra note
+    */
+    populateCheckoutForm (addressInfo) {
+        this.waitForComponent();
+        this.fields.mobileNumber.input.setValue(addressInfo.mobileNumber);
+        this.fields.addressLine1.input.setValue(addressInfo.line1);
+        this.fields.addressLine2.input.setValue(addressInfo.line2);
+        this.fields.addressCity.input.setValue(addressInfo.city);
+        this.fields.addressPostcode.input.setValue(addressInfo.postcode);
+        this.fields.userNote.input.setValue(addressInfo.note);
+    }
 
-/**
- * @description
- * Inputs address details into the checkout component.
- *
- * @param {Object} addressInfo
- * @param {String} addressInfo.mobileNumber The user's mobile number
- * @param {String} addressInfo.line1 First line of the user's address
- * @param {String} addressInfo.line2 Second line of the user's address
- * @param {String} addressInfo.city City of the user's address
- * @param {String} addressInfo.postcode Postcode of the user's address
- * @param {String} addressInfo.note The user's extra note
- */
-exports.populateCheckoutForm = addressInfo => {
-    exports.waitForCheckoutComponent();
-    fields.mobileNumber.input().setValue(addressInfo.mobileNumber);
-    fields.addressLine1.input().setValue(addressInfo.line1);
-    fields.addressLine2.input().setValue(addressInfo.line2);
-    fields.addressCity.input().setValue(addressInfo.city);
-    fields.addressPostcode.input().setValue(addressInfo.postcode);
-    fields.userNote.input().setValue(addressInfo.note);
-};
+    /**
+    * @description
+    * Changes checkout page to reflect checkout method to either delivery or collection depending on index given.
+    *
+    * @param {string} method The collection type: either 'delivery' or 'collection'
+    */
 
-/**
- * @description
- * Adds a space and backspaces to clear the value of the form field.
- * For more information on this, you can check out the link below:
- * https://github.com/webdriverio/webdriverio/issues/530#issuecomment-229435909
- *
- * @param {String} fieldName The name of the field input it is clearing
- */
-exports.clearField = fieldName => {
-    const BACKSPACE_UNICODE = '\uE003';
-    fields[fieldName].input().setValue([' ', BACKSPACE_UNICODE]);
-};
+    changeCheckoutMethod (method) {
+        const file = `/checkout-${method}.json`;
+        this.knobButton.click();
+        this.knobCheckoutDropdown.selectByVisibleText(file);
+    }
 
-/**
- * @description
- * Adds a space and backspaces to clear the value of the form field.
- *
- * @param {String} fields Grabs the fields of the above object and runs a forEach loop to get the keys
- */
-exports.clearCheckoutForm = fieldName => {
-    exports.waitForCheckoutComponent();
-    exports.clearField(fieldName);
-};
 
-exports.populateCollectionCheckoutForm = addressInfo => {
-    exports.waitForCheckoutComponent();
-    fields.mobileNumber.input().setValue(addressInfo.mobileNumber);
-    fields.userNote.input().setValue(addressInfo.note);
-};
+    /**
+    * @description
+    * Adds a space and backspaces to clear the value of the form field.
+    * For more information on this, you can check out the link below:
+    * https://github.com/webdriverio/webdriverio/issues/530#issuecomment-229435909
+    *
+    * @param {String} fieldName The name of the field input it is clearing
+    */
+    clearField (fieldName) {
+        const BACKSPACE_UNICODE = '\uE003';
+        this.fields[fieldName].input.setValue([' ', BACKSPACE_UNICODE]);
+    }
 
-/**
- * @description
- * Sets the value of the order time in dropdown based on visible text.
- *
- * @param {String} orderTime The visible text value of the order time
- */
-exports.selectOrderTime = orderTime => {
-    orderTimeDropdown().selectByVisibleText(orderTime);
-};
+    /**
+     * @description
+     * Adds a space and backspaces to clear the value of the form field.
+     *
+     * @param {String} fields Grabs the fields of the above object and runs a forEach loop to get the keys
+     */
+    clearCheckoutForm (fieldName) {
+        this.waitForComponent();
+        this.clearField(fieldName);
+    }
 
-/**
- * @description
- * The time of the order should increase when a higher index is applied.
- *
- * @param {Number} index The index of the `orderTimeDropdownOptions` array
- */
-exports.getOrderTimeOptionText = index => orderTimeDropdownOptions()[index].getText();
+    populateCollectionCheckoutForm (addressInfo) {
+        this.waitForComponent();
+        this.fields.mobileNumber.input.setValue(addressInfo.mobileNumber);
+        this.fields.userNote.input.setValue(addressInfo.note);
+    }
 
-/**
- * @description
- * Sets the value of the user note.
- *
- * @param {Object} addressInfo
- * @param {String} addressInfo.note The user's extra note
- */
-exports.inputUserNote = addressInfo => {
-    fields.userNote.input().setValue(addressInfo.note);
-};
+    /**
+    * @description
+    * Sets the value of the order time in dropdown based on visible text.
+    *
+    * @param {String} orderTime The visible text value of the order time
+    */
+    selectOrderTime (orderTime) {
+        this.orderTimeDropdown.selectByVisibleText(orderTime);
+    }
 
-/**
- * @description
- * Grabs the length of characters of the user note.
- *
- * @returns {number} The length of the user note
- */
-exports.getUserNoteLength = () => userNoteInput().getValue().length;
+    /**
+    * @description
+    * The time of the order should increase when a higher index is applied.
+    *
+    * @param {Number} index The index of the `orderTimeDropdownOptions` array
+    */
+    getOrderTimeOptionText (index) {
+        return this.orderTimeDropdownOptions[index].getText();
+    }
 
-/**
- * @description
- *Submit the checkout form.
- */
-exports.goToPayment = () => {
-    goToPaymentButton().click();
-};
+    /**
+     * @description
+     * Sets the value of the user note.
+     *
+     * @param {Object} addressInfo
+     * @param {String} addressInfo.note The user's extra note
+     */
+    inputUserNote (addressInfo) {
+        this.fields.userNote.input.setValue(addressInfo.note);
+    }
 
-exports.doesErrorMessageExist = errorMessage => doesElementExist(FIELDS[errorMessage].error);
-exports.doesFieldExist = inputField => doesElementExist(FIELDS[inputField].input);
+    /**
+    * @description
+    * Grabs the length of characters of the user note.
+    *
+    * @returns {number} The length of the user note
+    */
+    getUserNoteLength () {
+        return this.userNoteInput.getValue().length;
+    }
+
+    /**
+    * @description
+    *Submit the checkout form.
+    */
+    goToPayment () {
+        this.goToPaymentButton.click();
+    }
+
+    doesErrorMessageExist (errorMessage) {
+        return this.fields[errorMessage].error.isExisting();
+    }
+
+    doesFieldExist (inputField) {
+        return this.fields[inputField].input.isExisting();
+    }
+}
+
+module.exports = Checkout;