+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout-guest.component.spec.js
@@ -1,36 +1,45 @@
 import forEach from 'mocha-each';
-import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
+
+const Checkout = require('../../../test-utils/component-objects/f-checkout.component');
+
+const checkout = new Checkout();
 
 describe('f-checkout component tests', () => {
-    before(() => {
-        browser.url('iframe.html?id=components-organisms--checkout-component&knob-Get%20Checkout%20Url=%2Fcheckout-delivery.json&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json&knob-Create%20Guest%20Url=%2Fcreate-guest.json&knob-Get%20Basket%20Url=%2Fget-basket-delivery.json&knob-Auth%20token=&knob-Login%20Url=%2Flogin&viewMode=story');
-        CheckoutComponent.waitForCheckoutComponent();
+    beforeEach(() => {
+        const checkoutData = {
+            type: 'delivery',
+            isAuthenticated: false,
+            isValid: true
+        };
+
+        checkout.open(checkoutData);
+        checkout.waitForComponent();
     });
 
     it('should display the guest checkout header component', () => {
         // Assert
-        expect(CheckoutComponent.isGuestCheckoutHeaderDisplayed()).toBe(true);
+        expect(checkout.isGuestCheckoutHeaderDisplayed()).toBe(true);
     });
 
     it('should display the guest checkout login button', () => {
         // Assert
-        expect(CheckoutComponent.isGuestCheckoutLoginButtonDisplayed()).toBe(true);
+        expect(checkout.isGuestCheckoutLoginButtonDisplayed()).toBe(true);
     });
 
     forEach(['firstName', 'lastName', 'emailAddress'])
     .it('should show the guest checkout fields', field => {
         // Assert
-        expect(CheckoutComponent.doesFieldExist(field)).toBe(true);
+        expect(checkout.doesFieldExist(field)).toBe(true);
     });
 
     forEach(['firstName', 'lastName', 'emailAddress'])
     .it('should display each fields error message', field => {
         // Act
-        CheckoutComponent.clearCheckoutForm(field);
-        CheckoutComponent.goToPayment();
+        checkout.clearCheckoutForm(field);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed(field)).toBe(true);
+        expect(checkout.isFieldErrorDisplayed(field)).toBe(true);
     });
 
     it('should prevent user from submitting an invalid email address', () => {
@@ -40,23 +49,75 @@ describe('f-checkout component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(emailAddress);
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(emailAddress);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed('emailAddress')).toBe(true);
+        expect(checkout.isFieldErrorDisplayed('emailAddress')).toBe(true);
     });
-
 
     it('should navigate to correct url when the login link is clicked', () => {
         // Arrange
         const loginPath = '/login';
 
         // Act
-        CheckoutComponent.clickGuestCheckoutLogin();
+        checkout.clickGuestCheckoutLoginButton();
         const { pathname } = new URL(browser.getUrl());
 
         // Assert
         expect(pathname).toEqual(loginPath);
+    });
+
+    it('should display times in ascending order, with default text "As soon as possible" showing first', () => {
+        // Act
+        checkout.selectOrderTime('As soon as possible');
+
+        // Assert
+        expect(checkout.isOrderTimeDropdownDisplayed()).toBe(true);
+        expect(checkout.getOrderTimeOptionText(0)).toBe('As soon as possible');
+        expect(checkout.getOrderTimeOptionText(1)).toBe('Wednesday 00:45');
+        expect(checkout.getOrderTimeOptionText(2)).toBe('Wednesday 01:00');
+    });
+
+    it('should display a "mobileNumber" error message when an unsupported country code is used in the mobile number field', () => {
+        // Arrange
+        const addressDetails = {
+            mobileNumber: '+8112345678911'
+        };
+
+        // Act
+        checkout.populateCheckoutForm(addressDetails);
+        checkout.goToPayment();
+
+        // Assert
+        expect(checkout.isFieldErrorDisplayed('mobileNumber')).toBe(true);
+    });
+
+    it('should not display a "mobileNumber" error message when a number is formatted with a supported country code', () => {
+        // Arrange
+        const addressDetails = {
+            mobileNumber: '+4412345678911'
+        };
+
+        // Act
+        checkout.populateCheckoutForm(addressDetails);
+        checkout.goToPayment();
+
+        // Assert
+        expect(checkout.isFieldErrorDisplayed('mobileNumber')).toBe(false);
+    });
+
+    it('should prevent a user from writing a note of over 200 characters', () => {
+        // Arrange
+        const userNote = 'A';
+        const addressInfo = {
+            note: userNote.repeat(300)
+        };
+
+        // Act
+        checkout.inputUserNote(addressInfo);
+
+        // Assert
+        expect(checkout.userNoteMaxCharacterCount()).toEqual('200');
     });
 });